// src/types/constants.ts
export enum camera {
    ALL = "all",
    FHAZ = "fhaz",
    RHAZ = "rhaz",
    MAST = "mast",
    CHEMCAM = "chemcam",
    MAHLI = "mahli",
    MARDI = "mardi",
    NAVCAM = "navcam",
    PANCAM = "pancam",
    MINITES = "minites",
  }

	export enum viewOptions {
		GRID = "grid",
		SLIDER = "slider"
	}

	export enum ACTIONS {
		SET_PHOTO_DATA = "SET_PHOTO_DATA",
		SET_CAMERA = "SET_CAMERA",
		SET_SOL = "SET_SOL",
		SET_EARTH_DATE = "SET_EARTH_DATE",
		SET_EARTH_YEAR = "SET_EARTH_YEAR",
		SET_EARTH_MONTH = "SET_EARTH_MONTH",
		SET_EARTH_DAY = "SET_EARTH_DAY",
		SET_ROVER_TYPE = "SET_ROVER_TYPE"
	}